LayoutTests/media/picture-in-picture/picture-in-picture-api-enter-pip-1-expected.txt

This tests that request Picture-in-Picture requires a user gesture.

RUN(internals.settings.setAllowsPictureInPictureMediaPlayback(true))
RUN(internals.setMockVideoPresentationModeEnabled(true))
RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplaythrough)
EXPECTED (error.name == 'NotAllowedError') OK
END OF TEST
